requirements.txt
absl-py==0.9.0
natsort
amqp==2.4.2
celery==4.2.1
chardet==3.0.4
cloudpickle==1.3.0
cryptography==2.1.4
cycler==0.10.0
decorator==4.4.2
Distance==0.1.3
enum34==1.1.10
funcsigs==1.0.2
gast==0.3.3
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.12.0
google-auth-httplib2==0.0.3
google-pasta==0.2.0
googleapis-common-protos==1.51.0
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
html5lib==0.9999999
idna==2.6
ipaddress==1.0.17
ipython==5.5.0
ipython-genutils==0.2.0
joblib==0.14.1
matplotlib==2.2.5
mock==3.0.5
networkx==2.2
numpy==1.16.6
oauth2client==4.1.3
olefile==0.46
opencv-python==4.2.0.32
pandas==0.24.2
parsel==1.5.2
pathlib2==2.3.0
pbr==5.4.4
pep8==1.7.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==6.2.2
prompt-toolkit==1.0.15
pycairo==1.16.2
pycodestyle==2.3.1
pycrypto==2.6.1
pygame==1.9.6
Pygments==2.2.0
pytesseract
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.0.3
pyxdg==0.25
PyYAML==3.12
scandir==1.7
scikit-image==0.14.5
scipy==1.2.3
SecretStorage==2.3.1
selenium==3.141.0
subprocess32==3.5.4
tqdm==4.43.0
vine==1.2.0
wcwidth==0.1.7
Werkzeug==1.0.0
wrapt==1.12.1
